requirements.txt
pandas>=2.1,<3
nba_api>=1.5,<2
google-cloud-storage>=2.16,<3
google-cloud-bigquery>=3.25,<4
python-dotenv>=1.0,<2
pyarrow>=15,<20
pytest>=8.2,<9
dbt-bigquery>=1.8,<2
fastapi>=0.115,<1
uvicorn>=0.30,<1
jinja2>=3.1,<4
httpx>=0.27,<1

# Airflow (local development only; pre-installed in Cloud Composer)
apache-airflow>=2.7,<3
